[드로잉 앱] 스티치-repository 추가로 MVP 패턴 수정/ 우측 투명 seekBar 수정 시 사각형 실시간 반영 처리

+++ app/src/main/java/com/example/kotlindrawingapp/square/Point.kt
@@ -1,3 +1,3 @@
 package com.example.kotlindrawingapp.square
 
-data class Point(val x: Int, val y: Int) {}
+data class Point(val x: Float, val y: Float) {}
